+++ packages/core/src/wallets/index.ts
@@ -1,1 +1,2 @@
+export * from "./cosmostation";
 export * from "./keplr";